requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.25.2
python-dotenv==1.0.0
groq==0.4.1
requests==2.31.0
sentence-transformers==2.2.2
torch==2.1.1
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==1.1.0
